test: validate ported solc cases
Restore attributed tests to match their upstream solc programs, split unrelated and overmerged cases, and remove duplicate attributions. Reorganize codegen fixtures into focused directories and make MIR fixture discovery recursive.\n\nFix semantic gaps exposed by the restored cases in constant evaluation, overload selection, lvalue handling, operator validation, and qualified base calls.

+++ tests/ui/parser/using_operator_not_user_definable.sol
@@ -1,3 +1,5 @@
 // ported-from: test/libsolidity/syntaxTests/operators/userDefined/operator_parsing_non_user_definable.sol
 
-using {f as new} for uint; //~ ERROR: expected
+// The parser rejects the first non-user-definable operator and does not recover
+// within the directive, so later upstream variants are unreachable here.
+using {f as new} for uint256; //~ ERROR: expected
